refactor text parser
