+++ src/shared/user.tsx
@@ -440,9 +440,9 @@ export const User = memo(function User({
 
   if (status === 'pending') {
     return (
-      <div className="flex items-center gap-3">
+      <div className="flex items-start gap-3">
         <div className="h-10 w-10 shrink-0 animate-pulse rounded-lg bg-neutral-300 dark:bg-neutral-700" />
-        <div className="flex h-6 flex-1 items-start gap-2">
+        <div className="h-6 flex-1">
           <div className="h-4 w-36 animate-pulse rounded bg-neutral-300 dark:bg-neutral-700" />
         </div>
       </div>
