+++ src/utils/types.d.ts
@@ -1,8 +1,9 @@
 import { type NDKEvent, type NDKUserProfile } from '@nostr-dev-kit/ndk';
 import { type Response } from '@tauri-apps/plugin-http';
+import { ReactNode } from 'react';
 
 export interface RichContent {
-  parsed: string;
+  parsed: string | ReactNode[];
   images: string[];
   videos: string[];
   links: string[];
